feat: add massa-api-exports module
