+++ massa-api-exports/Cargo.toml
@@ -1,29 +1,17 @@
 [package]
-name = "massa_models"
+name = "massa_api_exports"
 version = "0.1.0"
 authors = ["Massa Labs <info@massa.net>"]
 edition = "2021"
 
 [dependencies]
-displaydoc = "0.2"
-lazy_static = "1.4"
-num_enum = "0.5"
-rust_decimal = "1.26"
 serde = { version = "1.0", features = ["derive"] }
-serde_with = "2.1.0"
-thiserror = "1.0"
-num = { version = "0.4", features = ["serde"] }
-directories = "4.0"
-config = "0.13"
-bs58 = { version = "0.4", features = ["check"] }
-bitvec = { version = "1.0", features = ["serde"] }
-nom = "7.1"
+strum = { version = "0.24", features = ["derive"] }
 
 # custom modules
-massa_hash = { path = "../massa-hash" }
-massa_serialization = { path = "../massa-serialization" }
 massa_signature = { path = "../massa-signature" }
 massa_time = { path = "../massa-time" }
+massa_models = { path = "../massa-models" }
 
 [dev-dependencies]
 serial_test = "0.10"
