+++ massa-api/src/public.rs
@@ -6,20 +6,19 @@ use crate::error::ApiError;
 use crate::{MassaRpcServer, Public, RpcServer, StopHandle, Value, API};
 use async_trait::async_trait;
 use jsonrpsee::core::{Error as JsonRpseeError, RpcResult};
+use massa_api_exports::{
+    DatastoreEntryInput, DatastoreEntryOutput, OperationInput, ReadOnlyBytecodeExecution,
+    ReadOnlyCall, ReadOnlyResult, SlotAmount,
+};
 use massa_consensus_exports::block_status::DiscardReason;
 use massa_consensus_exports::ConsensusController;
 use massa_execution_exports::{
     ExecutionController, ExecutionStackElement, ReadOnlyExecutionRequest, ReadOnlyExecutionTarget,
 };
-use massa_models::api::{
-    BlockGraphStatus, DatastoreEntryInput, DatastoreEntryOutput, OperationInput,
-    ReadOnlyBytecodeExecution, ReadOnlyCall, SlotAmount,
-};
-use massa_models::execution::ReadOnlyResult;
 use massa_models::operation::OperationDeserializer;
 use massa_models::secure_share::SecureShareDeserializer;
 use massa_models::{
-    block::Block, endorsement::SecureShareEndorsement, error::ModelsError,
+    api::BlockGraphStatus, block::Block, endorsement::SecureShareEndorsement, error::ModelsError,
     operation::SecureShareOperation, timeslots,
 };
 use massa_pos_exports::SelectorController;
@@ -27,19 +26,19 @@ use massa_protocol_exports::ProtocolCommandSender;
 use massa_serialization::{DeserializeError, Deserializer};
 
 use itertools::{izip, Itertools};
+use massa_api_exports::{
+    AddressInfo, BlockInfo, BlockInfoContent, BlockSummary, EndorsementInfo,
+    ExecuteReadOnlyResponse, NodeStatus, OperationInfo, TimeInterval,
+};
 use massa_models::datastore::DatastoreDeserializer;
 use massa_models::{
     address::Address,
-    api::{
-        AddressInfo, BlockInfo, BlockInfoContent, BlockSummary, EndorsementInfo, EventFilter,
-        NodeStatus, OperationInfo, TimeInterval,
-    },
+    api::EventFilter,
     block::BlockId,
     clique::Clique,
     composite::PubkeySig,
     config::CompactConfig,
     endorsement::EndorsementId,
-    execution::ExecuteReadOnlyResponse,
     node::NodeId,
     operation::OperationId,
     output_event::SCOutputEvent,
