+++ massa-client/Cargo.toml
@@ -22,6 +22,7 @@ strum = "0.24"
 strum_macros = "0.24"
 tokio = { version = "1.23", features = ["full"] }
 # custom modules
+massa_api_exports = { path = "../massa-api-exports" }
 massa_models = { path = "../massa-models" }
 massa_signature = { path = "../massa-signature" }
 massa_time = { path = "../massa-time" }
