+++ massa-sdk/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 [dependencies]
 jsonrpsee = { version = "0.16.2", features = ["client"] }
 http = "0.2.8"
+massa_api_exports = { path = "../massa-api-exports" }
 massa_models = { path = "../massa-models" }
 massa_time = { path = "../massa-time" }
